Moved types tests into separate folder
