+++ autoscheduler/frontend/src/tests/types/Meeting.test.ts
@@ -1,7 +1,7 @@
-import Meeting, { MeetingType } from '../types/Meeting';
-import { Indexable, fetchMock } from './util';
-import Section from '../types/Section';
-import Instructor from '../types/Instructor';
+import Meeting, { MeetingType } from '../../types/Meeting';
+import { Indexable, fetchMock } from '../util';
+import Section from '../../types/Section';
+import Instructor from '../../types/Instructor';
 
 // arrange
 const correctArgs: Indexable = {
